refactor: Address PR feedback on auth switch implementation
Addresses code review feedback from PR #1076:

1. **Test file organization** - Moved inline tests to separate
   rpc_auth_tests.rs file per CLAUDE.md convention for cleaner PR diffs

2. **Fixed misleading comment** - Removed comment that incorrectly
   implied an ordering bug. The real issue was the missing
   config.auth_provider assignment, not computation order

3. **Completed documentation** - Updated comment to document BOTH
   sources of wrong provider:
   - Reviewers: Get Claude from LaunchConfig::reviewer()
   - Non-reviewers: Get old provider from build_coworker_relaunch_config()

No functional changes - all documentation and organization improvements.
